Add benchmark and separate lib/main
